requirements.txt
absl-py==0.14.1
aniso8601==9.0.1
attrs==22.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
einops==0.3.2
flasgger==0.9.5
Flask==2.1.2
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
gevent==21.12.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
idna==3.3
imgviz==1.4.1
importlib-metadata==4.8.1
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
kiwisolver==1.3.2
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.2.2
mistune==2.0.4
numpy==1.21.4
oauthlib==3.1.1
onnx==1.12.0
onnxsim==0.4.6
opencv-python==4.5.3.56
pandas==1.3.4
Pillow==8.3.2
pkgutil_resolve_name==1.3.10
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyExecJS==1.5.1
Pygments==2.12.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
QtPy==1.11.1
requests==2.26.0
requests-oauthlib==1.3.0
rich==12.5.1
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
typing_extensions==4.3.0
urllib3==1.26.7
vit-pytorch==0.21.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
